api/src/main/java/com/thoughtworks/uaisoccer/teams/TeamStore.java
package com.thoughtworks.uaisoccer.teams;

import org.hibernate.Session;
import org.hibernate.SessionFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@Transactional
public class TeamStore {

    @Autowired
    SessionFactory sessionFactory;

    public Session getSession() {
        return sessionFactory.getCurrentSession();
    }

    public Long create(Team team) {
        return (Long)getSession().save(team);
    }
}